backend/requirements.txt
fastapi[standard]
httpx
motor
python-dotenv
pymongo[srv]
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
